diffeq-rag/requirements.txt
# Core Dependencies
langchain>=0.2.0
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0

# Vector Database
pymilvus>=2.4.0
milvus-langchain>=0.1.0

# Embedding Models
sentence-transformers>=2.5.0
flagembedding>=1.2.0

# Document Processing
pymupdf>=1.24.0
unstructured>=0.12.0
python-pptx>=0.6.23
python-docx>=1.1.0

# Text Processing
nltk>=3.8.0
scipy>=1.11.0

# Evaluation
ragas>=0.1.0
deepeval>=0.21.0

# API & Server
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
tiktoken>=0.7.0
tenacity>=8.2.0
loguru>=0.7.0

# Optional: For better PDF parsing
pdfplumber>=0.10.0
